get_icons_token_lists/requirements.txt
requests
svglib
python-magic
Pillow